requirements.txt
numpy==1.1.3
scipy==1.0.0
tensorflow_gpu==1.9.0
toml==0.9.2
